Add validators and sanitizers for user data #257
-Preventing backend errors with queries
