+++ src/main/resources/assets/js/graphql/principal/ListUserItemsRequest.ts
@@ -9,12 +9,12 @@ import {Role} from '../../app/principal/Role';
 import {UserJson} from '../../app/principal/UserJson';
 import {GroupJson} from '../../app/principal/GroupJson';
 import {RoleJson} from '../../app/principal/RoleJson';
-import {ListGraphQlRequest} from '../ListGraphQlRequest';
 import {UserItemBucketAggregationJson} from '../aggregation/UserItemBucketAggregationJson';
 import {UserItemAggregationHelper} from '../aggregation/UserItemAggregationHelper';
 import {UserItemType} from '../../app/browse/UserItemType';
 import {IdProvider} from '../../app/principal/IdProvider';
 import {IdProviderJson} from '../../app/principal/IdProviderJson';
+import {ListItemsRequest} from './ListItemsRequest';
 
 export type ListUserItemsRequestResult = {
     total: number,
@@ -23,24 +23,12 @@ export type ListUserItemsRequestResult = {
 };
 
 export class ListUserItemsRequest
-    extends ListGraphQlRequest<any> {
+    extends ListItemsRequest {
 
     private types: UserItemType[];
-    private searchQuery: string;
-    private itemIds: string[];
 
     setTypes(types: UserItemType[]): ListUserItemsRequest {
         this.types = types;
-        return this;
-    }
-
-    setQuery(query: string): ListUserItemsRequest {
-        this.searchQuery = query;
-        return this;
-    }
-
-    setItems(items: string[]): ListUserItemsRequest {
-        this.itemIds = items;
         return this;
     }
 
@@ -49,14 +37,6 @@ export class ListUserItemsRequest
 
         if (this.types && this.types.length > 0) {
             vars['types'] = this.types.map(type => UserItemType[type]);
-        }
-
-        if (this.searchQuery) {
-            vars['query'] = this.searchQuery;
-        }
-
-        if (this.itemIds && this.itemIds.length > 0) {
-            vars['itemIds'] = this.itemIds;
         }
 
         return vars;
